Remove references to companion website
Authorized by 2016 change request to remove mention of the DACS companion website since TS-DACS decided to focus on educational offerings rather than creating the website.  A few references to the website were missed at that time. Deleted a reference to the website in the Digital Records section and changed the entry about Appendix D in the What's New in DACS table.