Create License Archival Description Section
DACS Change Request
Submitted by Alston Cobourn, Patrick Galligan

Brief description of the component of DACS to be changed as well as the proposed change.
We are writing to request the creation of a new required DACS element 8.1 called "License for Archival Description" in Chapter 8 that would proceed the existing value added “Description Control” element. Repositories would use this element to declare that they are making the finding aid itself available under a license, such as machine actionable Creative Commons licenses.  We propose the following text:

2. Justification for the proposed change. 
  As Merrilee Proffitt from OCLC discussed in a Society of American Archivist (SAA) panel at the 2016 conference, opening finding aids for reuse is beneficial to researchers, particularly those in the digital humanities, as it allows them to use the data about our collections in new and interesting ways, such as text mining.  As researchers develop new ways to interact with our collections and finding aids, cultural heritage institutions need standardized ways to tell researchers and future archivists the ways in which they can use finding aid data.  Placing licenses on our finding aids affords us another way to fulfill our mission to help researchers use the archival collections with which we are entrusted.  Additionally, promoting open licensing aligns with our professional value to promote open and transparent cultural heritage.
The SAA Technical Subcommittee on Encoded Archival Standards has added a <rightsdeclaration> element to its EAD3 1.1 release in hopes of accommodating this need.  This release is now out for community review.  I have submitted a new feature request to ArchivesSpace asking for the creation of a corresponding field in the Finding Aid Information section of Resource records.  Bringing both standards and this system that utilizes them into alignment on this point would both provide a guideline for how to implement licenses as well as the ability to do so in a standardized way.

3.  Impact of the proposed change.
Adding this DACS element would encourage archives to open their finding aids by enabling them to do so easily.  Since archives very clearly own the copyrights to their finding aids and generating Creative Commons licenses is simple, the lack of appropriate DACS and EAD3 elements in which to record the license information are the only real impediments to archives that wish to open their finding aids.  Clarifying the restrictions on how the data about our collections can be used also has the potential to increase the awareness of our collections and make archives relevant to new researchers and research methods.